Update rebuild script

+++ private/scripts/refresh-local.sh
@@ -7,9 +7,9 @@ terminus backup:create --element=database durham-atletico.live
 echo "Downloading backup"
 DB_URL=$(terminus backup:get --element=db durham-atletico.live)
 curl -L -o database.sql.gz $DB_URL
-rm database.sql
+if [ -f database.sql ]; then rm database.sql; fi
 # Drop local DB
-docker exec -i durhamatletico_php_1 drush sql-drop -y
+docker exec -i durhamatletico_php_1 drush -r /var/www/html/web sql-drop -y
 echo "y" | gunzip database.sql.gz
 echo "Importing backup"
 pv database.sql | docker exec -i durhamatletico_mariadb_1 mysql -uroot -proot durhamatletico_docker
